requirements.txt
fastapi
uvicorn
pydantic
langgraph
pandas
openpyxl
ortools
httpx
python-dotenv
python-multipart